Support limited file recognization (#12)
In some OSs or some versions of the `file` command line tool, Excel files are recognized as `Zip archive data, at least vN.N to extract`. When the files come as ActionDispatch::Http::UploadedFile then it has an uninteresting filename so the extension cannot be used anyways.

With this change, Office packaged files are recognized.